centralize log initialisation
